Add option to disable looking for Root package.json file
Automatic search for a package.json as root of a project made `deglob`
hard to use in custom directory structures with custom `.Xignore` files.
New option `opts.findRoot` defaults to true, as to not alter the
original behaviour of `deglob` and disturb the flow of projects already
using it.
Setting `opts.findRoot = false` will disable the search for the closest
`package.json`. Instead, the `root` variable will be set to the
directory specified in `opts.cwd`.